LayoutTests/fast/dom/shadow/content-shadow-unknown.html

<!DOCTYPE html>
<html>
<head>
<script src="../../js/resources/js-test-pre.js"></script>
</head>
<body>
<div>
  <content id="content-simplest"></content>
</div>
<script>
description("Shadow DOM related elements should be hidden if the feature is off.");
if (window.internals)
    internals.settings.setShadowDOMEnabled(false);
var content = document.createElement("content");
shouldBeTrue("0 <= content.constructor.toString().indexOf('HTMLUnknownElement')");
var shadow = document.createElement("shadow");
shouldBeTrue("0 <= shadow.constructor.toString().indexOf('HTMLUnknownElement')");
</script>
<script src="../../js/resources/js-test-post.js"></script>
</body>
</html>
