LayoutTests/fast/dom/shadow/content-shadow-unknown-expected.txt

Shadow DOM related elements should be hidden if the feature is off.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS 0 <= content.constructor.toString().indexOf('HTMLUnknownElement') is true
PASS 0 <= shadow.constructor.toString().indexOf('HTMLUnknownElement') is true
PASS successfullyParsed is true

TEST COMPLETE
